requirements.txt
attrs==23.2.0
certifi==2024.7.4
exceptiongroup==1.2.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
outcome==1.3.0.post0
playwright==1.45.0
pyee==11.1.0
PySocks==1.7.1
python-dotenv==1.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
